+++ src/test/java/com/palominolabs/metrics/guice/ExceptionMeteredTest.java
@@ -11,8 +11,8 @@ import org.junit.Test;
 
 import static com.codahale.metrics.MetricRegistry.name;
 import static com.codahale.metrics.annotation.ExceptionMetered.DEFAULT_NAME_SUFFIX;
-import static com.palominolabs.metrics.guice.DefaultMetricNamer.METERED_SUFFIX;
-import static com.palominolabs.metrics.guice.DefaultMetricNamer.TIMED_SUFFIX;
+import static com.palominolabs.metrics.guice.DeclaringClassMetricNamer.METERED_SUFFIX;
+import static com.palominolabs.metrics.guice.DeclaringClassMetricNamer.TIMED_SUFFIX;
 import static org.hamcrest.Matchers.instanceOf;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.notNullValue;