+++ src/test/java/com/palominolabs/metrics/guice/MeteredTest.java
@@ -8,7 +8,7 @@ import org.junit.Before;
 import org.junit.Test;
 
 import static com.codahale.metrics.MetricRegistry.name;
-import static com.palominolabs.metrics.guice.DefaultMetricNamer.METERED_SUFFIX;
+import static com.palominolabs.metrics.guice.DeclaringClassMetricNamer.METERED_SUFFIX;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.notNullValue;
 import static org.junit.Assert.assertThat;